requirements.txt
attrs==20.3.0
autopep8==1.5.6
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
coverage==5.5
dash==1.19.0
dash-auth==1.4.1
dash-bootstrap-components==0.12.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
flake8==3.9.0
Flask==1.1.2
Flask-Compress==1.9.0
Flask-SeaSurf==0.3.0
future==0.18.2
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
plotly==4.14.3
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.15.0
toml==0.10.2
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
